requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
psycopg[binary]
SQLAlchemy==2.0.30
pgvector==0.2.5
passlib==1.7.4
bcrypt==4.1.3
PyJWT==2.8.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-openai==0.1.22
langchain-experimental==0.0.64
tenacity==8.4.1
rank-bm25==0.2.2
sentence-transformers==3.0.1
PyPDF2==3.0.1
python-dotenv==1.0.1
